be-service-boimetrics-ali/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.32
python-multipart>=0.0.9,<0.1
pillow>=10.4,<11
numpy>=1.24,<2
pymongo[srv]>=4.8,<5
python-dotenv>=1.0,<2
pytest>=8.2,<9
